added input validation to POST methods.

+++ db/scavenge_schema.sql
@@ -76,7 +76,7 @@ CREATE TABLE items (
     id              serial,
     hunt_id         int NOT NULL,
     name            varchar(255) NOT NULL CHECK (length(name) > 0),
-    points          smallint DEFAULT 1 CHECK (points > 0),
+    points          int DEFAULT 1 CHECK (points > 0),
     CONSTRAINT items_in_same_hunt_name UNIQUE(hunt_id, name),
     PRIMARY KEY(id),
     FOREIGN KEY (hunt_id) REFERENCES hunts(id) ON DELETE CASCADE
